// src/app/models/viewModels/climatiq-search-models/estimate/selectorViewModel.ts
export class SelectorViewModel {
  id: string;
  data_version: string;
  activity_id: string;
  source: string;
  source_dataset: string;
  region: string;
  region_fallback: boolean;
  year_fallback: boolean;
  year: number;
  source_lca_activity: string;
  calculation_method: string;
  allowed_data_quality_flags: string;

}